[feat] 카테고리 변경 구현 #158
게임 설정 중 카테고리 변경 기능 구현
게임 방을 만들고 설정 모달 안띄우고 바로 시작하면 모든 카테고리를 포함하도록 하였음
