+++ front-end/src/components/Game/GameButtons.tsx
@@ -1,7 +1,8 @@
 import React, { useState, useContext } from 'react';
 import { useHistory } from 'react-router';
-import { useSetRecoilState } from 'recoil';
+import { useRecoilState, useSetRecoilState } from 'recoil';
 import { globalContext } from '../../App';
+import globalAtom from '../../recoilStore/globalAtom';
 import globalSelector from '../../recoilStore/globalSelector';
 import { GAME_MESSAGE, ROOM_MEESSAGE } from '../../utils/socketMsgConstants';
 import { modalPropsType } from '../public/Modal';
@@ -13,6 +14,7 @@ const GameButtons = (props: GameButtonsPropsType) => {
   const { isOwner, roomTitle } = props;
   const history = useHistory();
   const [settingsModal, setSettingsModal] = useState([]);
+  const [roomSettings, setRoomSettings] = useRecoilState(globalAtom.roomSettings);
   const { socket } = useContext(globalContext);
   const popModal: (modalProps: modalPropsType) => void = useSetRecoilState(globalSelector.popModal);
 
@@ -25,7 +27,7 @@ const GameButtons = (props: GameButtonsPropsType) => {
     history.replace('/lobby');
   };
 
-  const roomSettings = () => {
+  const roomSettingsUpdate = () => {
     const ModalOutLocation = <section className="modal-outter" onClick={offModal} key={0} />;
     setSettingsModal([ModalOutLocation, <GameRoomSettings offModal={offModal} key={1} />]);
   };
@@ -36,8 +38,11 @@ const GameButtons = (props: GameButtonsPropsType) => {
       return;
     }
 
-    const category: string[] = ['과일', '탈것', '장소', '직업'];
-
+    const category: string[] = roomSettings.category
+      .map(({ category, include }) => {
+        if (include) return category;
+      })
+      .filter((category) => category !== undefined);
     socket.emit(GAME_MESSAGE.WORD_SELECT, { category, roomTitle });
   };
 
@@ -49,7 +54,7 @@ const GameButtons = (props: GameButtonsPropsType) => {
     <div className="game-header-buttons">
       {isOwner && (
         <>
-          <button className="game-header-button" onClick={roomSettings}>
+          <button className="game-header-button" onClick={roomSettingsUpdate}>
             게임 설정
           </button>
           {settingsModal}
