+++ front-end/src/components/Game/GameRoomSettings.tsx
@@ -8,24 +8,12 @@ import downArorw from '../../images/downArrow.svg';
 import { globalContext } from '../../App';
 import { GAME_MESSAGE } from '../../utils/socketMsgConstants';
 
-const categoryList = [
-  { category: '과일', include: true },
-  { category: '탈 것', include: true },
-  { category: '장소', include: true },
-  { category: '직업', include: true },
-  { category: '동물', include: true },
-  { category: '음식', include: true },
-  { category: '나라', include: true },
-  { category: '악기', include: true },
-  { category: '스포츠', include: true },
-];
-
 const GameRoomSettings = ({ offModal }: { offModal(): void }) => {
   const { socket } = useContext(globalContext);
   const user = useRecoilValue(globalAtom.user);
   const [roomSettings, setRoomSettings] = useRecoilState(globalAtom.roomSettings);
   const [roomInfo, setRoomInfo] = useState({ max: roomSettings.max, cycle: roomSettings.cycle, owner: user.user_id });
-  const [categories, setCategory] = useState(roomSettings.category.length === 0 ? categoryList : roomSettings.category);
+  const [categories, setCategory] = useState(roomSettings.category);
   const { selectedRoomTitle } = useRecoilValue(globalAtom.roomData);
   const client = useRecoilValue(globalAtom.client);
